// src/app/models/projects.enums.ts
export enum Projects {
  TILE_COLOR = 'tileColor',
  GITHUB_SHOWCASE = 'githubShowcase',
  NOTE_READER = 'noteReader',
  SEARCH_BOOKS = 'searchBooks',
  SELECT_ALL_FORM = 'selectAllForm',
  TRAFFIC_LIGHT = 'trafficLight',
  CALENDAR = 'calendar',
  TODO_LIST = 'todoList',
  USER_CARD = 'userCard',
}